Update blueprints and update testing
